requirements.txt
# Core Dependencies for Modular Image Editor
# Python 3.8+

# Image processing
Pillow>=10.0.0

# Numerical operations
numpy>=1.24.0

# Visualization and plotting
matplotlib>=3.7.0

# Interactive notebooks
jupyter>=1.0.0

# Optional: for advanced metrics
# scikit-image>=0.21.0
# opencv-python>=4.8.0
